wip command refactoring
